pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-interview-env"
version = "0.1.0"
description = "OpenEnv-compliant adaptive CS technical interview environment for training a scorer LLM"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "transformers>=4.40",
    "torch>=2.1",
    "gradio>=4.0",
    "plotly>=5.0",
    "pyyaml>=6.0",
    "hypothesis>=6.0",
    "pytest>=7.0",
    "scipy>=1.10",
    # "openenv",  # add once package name is confirmed
    # "trl>=0.8",
    # "unsloth",
]

[project.optional-dependencies]
train = ["trl>=0.8", "wandb", "unsloth"]
dev = ["pytest", "hypothesis", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["adaptive_interview_env*"]
